backend/pyproject.toml
[project]
name = "syna-backend"
version = "0.1.0"
description = "AI Operating System for Conversation-Native Work"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.12",
    "httpx>=0.27.0",
    "redis>=5.1.0",
    "asyncpg>=0.30.0",
    "alembic>=1.13.0",
    "langchain>=0.3.0",
    "langgraph>=0.2.74",
    "langchain-openai>=0.2.0",
    "langchain-anthropic>=0.2.0",
    "langchain-google-genai>=2.0.0",
    "openai>=1.54.0",
    "anthropic>=0.39.0",
    "google-generativeai>=0.8.0",
    "pgvector>=0.3.0",
    "supabase>=2.10.0",
    "pyyaml>=6.0.2",
    "python-dotenv>=1.0.1",
    "sentry-sdk[fastapi]>=2.17.0",
    "prometheus-client>=0.21.0",
    "opentelemetry-api>=1.28.0",
    "opentelemetry-sdk>=1.28.0",
    "opentelemetry-instrumentation-fastapi>=0.48b0",
    "email-validator>=2.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "black>=24.10.0",
    "ruff>=0.7.0",
    "mypy>=1.13.0",
    "pre-commit>=4.0.0",
    "ipython>=8.29.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py311"
line-length = 88
select = ["E", "F", "I", "N", "W", "B", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
asyncio_mode = "auto"

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.hatch.build.targets.wheel]
packages = ["app"]
